requirements.txt
python-telegram-bot==21.5
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
sqlalchemy==2.0.36
